requirements.txt
autoflake==0.6.6
awesome-slugify==1.6.5
babelfish==0.5.5
decorator==4.0.10
guessit==2.0.5
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
langdetect==1.0.6
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.3
ptyprocess==0.5.1
pyflakes==1.2.3
Pygments==2.1.3
python-dateutil==2.5.1
rebulk==0.7.3
regex==2016.7.21
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
tinydb==3.2.1
tqdm==4.8.3
traitlets==4.2.2
Unidecode==0.4.19
wcwidth==0.1.7
